services/local-rag-ai-assistant/requirements.txt
# Core LangChain
langchain>=1.2,<2.0
langchain-community>=0.4,<1.0
langchain-ollama>=1.0,<2.0
langchain-qdrant>=1.1,<2.0
langchain-text-splitters>=1.0,<2.0

# Vector store & document loading
qdrant-client>=1.17,<2.0
pypdf>=4.0,<6.0

# Web framework
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
python-multipart>=0.0.9,<1.0
pydantic>=2.0,<3.0

# Ollama client
ollama>=0.4,<1.0

# Gradio UI (used by rag_assistant.py)
gradio>=5.0,<6.0

# Optional — loads .env files when present
python-dotenv>=1.0,<2.0
